#318: rename "onlinejudge.implementation" "onlinejudge._implementation"
