--- docs/onlinejudge.implementation.command.code_statistics.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.code\_statistics module
-==========================================================
-
-.. automodule:: onlinejudge.implementation.command.code_statistics
-    :members:
-    :undoc-members:
-    :show-inheritance:
