--- docs/onlinejudge.implementation.command.download.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.download module
-==================================================
-
-.. automodule:: onlinejudge.implementation.command.download
-    :members:
-    :undoc-members:
-    :show-inheritance:
