--- docs/onlinejudge.implementation.command.generate_output.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.generate\_output module
-==========================================================
-
-.. automodule:: onlinejudge.implementation.command.generate_output
-    :members:
-    :undoc-members:
-    :show-inheritance:
