--- docs/onlinejudge.implementation.command.generate_scanner.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.generate\_scanner module
-===========================================================
-
-.. automodule:: onlinejudge.implementation.command.generate_scanner
-    :members:
-    :undoc-members:
-    :show-inheritance:
