--- docs/onlinejudge.implementation.command.get_standings.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.get\_standings module
-========================================================
-
-.. automodule:: onlinejudge.implementation.command.get_standings
-    :members:
-    :undoc-members:
-    :show-inheritance:
