--- docs/onlinejudge.implementation.command.login.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.login module
-===============================================
-
-.. automodule:: onlinejudge.implementation.command.login
-    :members:
-    :undoc-members:
-    :show-inheritance:
