--- docs/onlinejudge.implementation.command.rst
@@ -1,26 +1,0 @@
-onlinejudge.implementation.command package
-==========================================
-
-Submodules
-----------
-
-.. toctree::
-
-   onlinejudge.implementation.command.code_statistics
-   onlinejudge.implementation.command.download
-   onlinejudge.implementation.command.generate_output
-   onlinejudge.implementation.command.generate_scanner
-   onlinejudge.implementation.command.get_standings
-   onlinejudge.implementation.command.login
-   onlinejudge.implementation.command.split_input
-   onlinejudge.implementation.command.submit
-   onlinejudge.implementation.command.test
-   onlinejudge.implementation.command.test_reactive
-
-Module contents
----------------
-
-.. automodule:: onlinejudge.implementation.command
-    :members:
-    :undoc-members:
-    :show-inheritance:
