--- docs/onlinejudge.implementation.command.split_input.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.split\_input module
-======================================================
-
-.. automodule:: onlinejudge.implementation.command.split_input
-    :members:
-    :undoc-members:
-    :show-inheritance:
