--- docs/onlinejudge.implementation.command.submit.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.submit module
-================================================
-
-.. automodule:: onlinejudge.implementation.command.submit
-    :members:
-    :undoc-members:
-    :show-inheritance:
