--- docs/onlinejudge.implementation.command.test.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.test module
-==============================================
-
-.. automodule:: onlinejudge.implementation.command.test
-    :members:
-    :undoc-members:
-    :show-inheritance:
