--- docs/onlinejudge.implementation.command.test_reactive.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.command.test\_reactive module
-========================================================
-
-.. automodule:: onlinejudge.implementation.command.test_reactive
-    :members:
-    :undoc-members:
-    :show-inheritance:
