--- docs/onlinejudge.implementation.download_history.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.download\_history module
-===================================================
-
-.. automodule:: onlinejudge.implementation.download_history
-    :members:
-    :undoc-members:
-    :show-inheritance:
