--- docs/onlinejudge.implementation.format_utils.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.format\_utils module
-===============================================
-
-.. automodule:: onlinejudge.implementation.format_utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
