--- docs/onlinejudge.implementation.logging.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.logging module
-=========================================
-
-.. automodule:: onlinejudge.implementation.logging
-    :members:
-    :undoc-members:
-    :show-inheritance:
