--- docs/onlinejudge.implementation.main.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.main module
-======================================
-
-.. automodule:: onlinejudge.implementation.main
-    :members:
-    :undoc-members:
-    :show-inheritance:
