--- docs/onlinejudge.implementation.rst
@@ -1,28 +1,0 @@
-onlinejudge.implementation package
-==================================
-
-Subpackages
------------
-
-.. toctree::
-
-    onlinejudge.implementation.command
-
-Submodules
-----------
-
-.. toctree::
-
-   onlinejudge.implementation.download_history
-   onlinejudge.implementation.format_utils
-   onlinejudge.implementation.logging
-   onlinejudge.implementation.main
-   onlinejudge.implementation.utils
-
-Module contents
----------------
-
-.. automodule:: onlinejudge.implementation
-    :members:
-    :undoc-members:
-    :show-inheritance:
