--- docs/onlinejudge.implementation.utils.rst
@@ -1,7 +1,0 @@
-onlinejudge.implementation.utils module
-=======================================
-
-.. automodule:: onlinejudge.implementation.utils
-    :members:
-    :undoc-members:
-    :show-inheritance:
