+++ oj
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import sys
 
-import onlinejudge.implementation.main
+import onlinejudge._implementation.main
 
 if __name__ == '__main__':
-    onlinejudge.implementation.main.main(args=sys.argv[1:])
+    onlinejudge._implementation.main.main(args=sys.argv[1:])
