+++ onlinejudge/_implementation/command/code_statistics.py
@@ -7,8 +7,8 @@ import subprocess
 from typing import *
 
 import onlinejudge
-import onlinejudge.implementation.logging as log
-import onlinejudge.implementation.utils as utils
+import onlinejudge._implementation.logging as log
+import onlinejudge._implementation.utils as utils
 
 if TYPE_CHECKING:
     import argparse
