+++ onlinejudge/_implementation/command/get_standings.py
@@ -4,8 +4,8 @@ import sys
 from typing import *
 
 import onlinejudge
-import onlinejudge.implementation.logging as log
-import onlinejudge.implementation.utils as utils
+import onlinejudge._implementation.logging as log
+import onlinejudge._implementation.utils as utils
 
 if TYPE_CHECKING:
     import argparse
