+++ onlinejudge/_implementation/command/split_input.py
@@ -6,8 +6,8 @@ from typing import *
 from typing.io import *
 
 import onlinejudge
-import onlinejudge.implementation.logging as log
-import onlinejudge.implementation.utils as utils
+import onlinejudge._implementation.logging as log
+import onlinejudge._implementation.utils as utils
 
 if TYPE_CHECKING:
     import argparse
