+++ onlinejudge/_implementation/utils.py
@@ -22,7 +22,7 @@ import bs4
 import requests
 
 import onlinejudge.__about__ as version
-import onlinejudge.implementation.logging as log
+import onlinejudge._implementation.logging as log
 from onlinejudge.type import LabeledString, TestCase
 
 config_dir = pathlib.Path(appdirs.user_config_dir(version.__package_name__))
