+++ onlinejudge/dispatch.py
@@ -21,8 +21,8 @@
 
 from typing import TYPE_CHECKING, List, Optional, Type
 
-import onlinejudge.implementation.logging as log
-from onlinejudge.type import Service, Problem, Submission
+import onlinejudge._implementation.logging as log
+from onlinejudge.type import Problem, Service, Submission
 
 submissions = []  # type: List[Type['Submission']]
