+++ onlinejudge/service/aoj.py
@@ -12,9 +12,9 @@ from typing import *
 
 import requests
 
+import onlinejudge._implementation.logging as log
+import onlinejudge._implementation.utils as utils
 import onlinejudge.dispatch
-import onlinejudge.implementation.logging as log
-import onlinejudge.implementation.utils as utils
 import onlinejudge.type
 from onlinejudge.type import LabeledString, TestCase
