+++ onlinejudge/service/atcoder.py
@@ -11,9 +11,9 @@ from typing import *
 import bs4
 import requests
 
+import onlinejudge._implementation.logging as log
+import onlinejudge._implementation.utils as utils
 import onlinejudge.dispatch
-import onlinejudge.implementation.logging as log
-import onlinejudge.implementation.utils as utils
 import onlinejudge.type
 from onlinejudge.type import SubmissionError
